+++ release_notes.md
@@ -15,9 +15,11 @@ PREESM Changelog
 * Force Graphiti to [1.9.0,2.0.0);
 * Force DFTools to [1.8.0,2.0.0);
 * Update Xtend/Xtext to 2.14.0+;
+* Fix papify codegen for spider
 
 ### Bug fix
 * Fix issue #61;
+
 
 ## Release version 2.12.3
 *2018.06.06*
@@ -50,6 +52,7 @@ PREESM Changelog
 * In spider codegen, the try - catch in the main auto generated now englobe the init as well;
 * Update manifest files;
 * Fix xtend warnings;
+* Adding papify code generation for spider
 
 ### Bug fix
 * Fix issue with copy/paste when reconnecting dependencies;